Gemify the code into an new internal gem 'spree_frontend_bootstrap'.
spree_bootstrap_frontend currently contains overrides for different parts
of spree and spree_auth_devise. This causes some problems for pushing
these changes back to the upstream projects. This is especially true
for the spree_auth_devise changes as spree_auth_devise is a separate
repo and requires a whole other PR/fork.

This is the start of the work to split the overrides for each project.

For more details see https://github.com/spree/spree/pull/5044.