Lock Rails to 4.1.4.
The checkout specs are failing because versions of Rails later than
4.1.4 upgrade the mail gem. Spree did use the internal API of the mail
gem to verify email addresses but this internal API has been changed
in later versions of mail and the changes break Spree's email validation.

Spree has fixed this problem but spree_bootstrap_frontend has not been
synced to that commit yet.